Make sure that dhandle structure doesn't grow in size.

+++ src/include/dhandle.h
@@ -150,8 +150,8 @@ struct __wt_data_handle {
 #define WT_DHANDLE_TS_MIXED_MODE 0x04u         /* Handle using mixed mode timestamps checking. */
 #define WT_DHANDLE_TS_NEVER 0x08u              /* Handle never using timestamps checking. */
 #define WT_DHANDLE_TS_ORDERED 0x10u            /* Handle using ordered timestamps checking. */
-                                               /* AUTOMATIC FLAG VALUE GENERATION STOP 32 */
-    uint32_t ts_flags;
+                                               /* AUTOMATIC FLAG VALUE GENERATION STOP 16 */
+    uint16_t ts_flags;
 
 #define WT_DHANDLE_LOCK_WRITE 0x01u /* Write lock is acquired. */
     uint16_t lock_flags;
